pyproject.toml
[tool.pdm]
version = { source = "file", path = "src/sanic_book/version.py" }
allow_prereleases = true

[project]
name = "sanic-book"
dynamic = ["version"]
authors = [
    {name = "xinetzone", email = "xinzone@outlook.com"},
]
dependencies = [
  "d2py",
  "sanic",
  "strawberry-graphql[debug-server]",
  "importlib-metadata>=4.12"
]
requires-python = ">=3.10"
readme = "README.md"
license = {file = "LICENSE"}

[build-system]
requires = ["pdm-pep517"]
build-backend = "pdm.pep517.api"

[project.urls]
Home = "https://github.com/xinetzone/sanic-book"

# [project.entry-points]
# "sphinx.html_themes" = {tvm_book = "tvm_book"}

[project.optional-dependencies]
doc = [
  "xyzstyle",
  "myst-nb",
  "sphinx-graphql",
  # "pytest-regressions",
  # "sphinxext-rediraffe",
  "sphinx-copybutton",
  # "sphinx-intl",
  "ablog",
  "sphinx-thebe",
  "sphinx-comments",
  "plotly",
  "pandas",
  "ipywidgets",
  # "pandas",
  # "matplotlib",
  "sphinx-design",
  "sphinxcontrib-bibtex",
  "sphinx-sitemap",
  "sphinx>=5.3"
]

[tool.pdm.build]
package-dir = "src"
# includes = []
# excludes = []
# source-includes = ["tests/"]

[pypi.extra]
url = "https://pypi.tuna.tsinghua.edu.cn/simple "
# verify_ssl = false